fix(install): keep the identity panel's imports out of the server graph
Production Build, UX Route Budget Sweep, and UX Route Sweep Runtime all
failed on one cause: `node:fs/promises` reached the CLIENT bundle for
/(shell)/workspace/page.

The panel is a "use client" component, so every module it imports for a
VALUE is bundled for the browser. It imported `CONFIRMATION_PRESENTATION`
from installation-identity-view, which reaches installer state through
node:fs/promises, and two label maps from identity-change-impact, which
hashes with node:crypto. Turbopack: "the chunking context does not
support external modules".

Split the contracts along the boundary that actually exists:

- `identity-presentation.ts` (new) — labels, option lists, and the shape
  of the data the panel renders. No value import that touches Node;
  @dpf/db/installation-operating-intent is imported for TYPES ONLY,
  because it pulls in `crypto` for its fingerprint helper.
- `environment-class-contract.ts` (new) — the pure precedence contract,
  split from the filesystem read that applies it. `environment-class.ts`
  re-exports it so server callers keep one import.
- `identity-change-impact.ts` and `installation-identity-view.ts` keep
  only what genuinely needs crypto and the filesystem.

The option lists are now spelled out rather than mapped from
INSTALLATION_OPERATING_PURPOSES, so the panel does not bundle a crypto
polyfill for two lists of strings. `identity-presentation.test.ts`
asserts they stay identical to the closed vocabularies, and asserts the
module takes no Node built-in and no value import from a server sibling
— so the regression cannot come back quietly.

Verified with a real `pnpm --filter web build` this time, which is what
typecheck and vitest could not see. `check-bundle-boundaries` passed
throughout: it checks direct imports in a "use client" file, not the
transitive graph.
